pyproject.toml
[project]
name = "pm-angel"
version = "0.1.0"
description = "Bot de copy-trading pour Polymarket"
requires-python = ">=3.12"
dependencies = [
    "py-clob-client>=0.34.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.20.0",
    "sqlalchemy[asyncio]>=2.0",
    "rich>=13.0.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
pm-angel = "pm_angel.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pm_angel"]
